# [community] Fix index so that it's used in uniqueness query.