--- a/ecosystem/platform/server/spec/factories/users.rb
+++ b/ecosystem/platform/server/spec/factories/users.rb
@@ -5,7 +5,7 @@
 
 FactoryBot.define do
   factory :user, class: User do
-    username { Faker::Internet.username(specifier: 3..20, separators: %w[- _]) }
+    username { Faker::Internet.username(specifier: 3..14, separators: %w[- _]) + 5.times.map { rand(10) }.join }
     password { Faker::Internet.password }
     email { Faker::Internet.email }
     bio { Faker::Lorem.paragraph }