### Description

The query to determine `username` uniqueness was not optimized.

Additionally, this PR adjusts the user factory to reduce the chance of duplicate usernames.

Before:

```
postgres=> explain analyze select 1 as one from "users" where lower("users"."username") = lower('zach') and "users"."id" != 9000000 limit 1;
                                                 QUERY PLAN
------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.00..10.87 rows=1 width=4) (actual time=37.787..37.789 rows=1 loops=1)
   ->  Seq Scan on users  (cost=0.00..9556.35 rows=879 width=4) (actual time=37.786..37.787 rows=1 loops=1)
         Filter: ((id <> 9000000) AND (lower((username)::text) = 'zach'::text))
         Rows Removed by Filter: 76098
 Planning Time: 0.873 ms
 Execution Time: 37.830 ms
(6 rows)
```

After:

```
community_platform_dev=# explain analyze select 1 as one from "users" where lower("users"."username") = lower('zach') and "users"."id" != 9000000 limit 1;
                                                                QUERY PLAN
-------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.14..8.16 rows=1 width=4) (actual time=0.009..0.010 rows=0 loops=1)
   ->  Index Scan using index_users_on_lower_username on users  (cost=0.14..8.16 rows=1 width=4) (actual time=0.009..0.009 rows=0 loops=1)
         Index Cond: (lower((username)::text) = 'zach'::text)
         Filter: (id <> 9000000)
 Planning Time: 0.372 ms
 Execution Time: 0.022 ms
(6 rows)
```

### Test Plan
Manual

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/aptos-labs/aptos-core/4257)
<!-- Reviewable:end -->